fixed EVERYTHING

+++ Arduino/libraries/sensor/sensor.cpp
@@ -17,7 +17,7 @@ uint8_t fifoBuffer[64]; // FIFO storage buffer
 
 Quaternion q;           // [w, x, y, z]         quaternion container
 VectorFloat gravity;    // [x, y, z]            gravity vector
-double ypr[3];           // [yaw, pitch, roll]   yaw/pitch/roll container and gravity vector
+float ypr[3];           // [yaw, pitch, roll]   yaw/pitch/roll container and gravity vector
 int16_t gx, gy, gz;
 volatile bool mpuInterrupt = false;     // indicates whether MPU interrupt pin has gone high
